pyproject.toml
[project]
name = "reports"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-genai>=1.65.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.1",
    "playwright>=1.58.0",
    "pyinstaller>=6.16.0",
    "python-dotenv>=1.2.2",
    "python-pptx>=1.0.2",
    "uvicorn[standard]>=0.41.0",
]
